+++ src/ipsec/isakmp/transform.rs
@@ -7,7 +7,7 @@ use super::attr::{AttributeIter};
 const KEY_IKE: u8 = 0x01;
 
 
-#[derive(Copy)]
+#[derive(Clone,Copy)]
 pub struct Transform<'a>
 {
     pub TransformNum: u8,
@@ -19,7 +19,7 @@ pub struct Transform<'a>
 
 impl<'a> Transform<'a>
 {
-    pub fn HeaderSize() -> uint { 2 }
+    pub fn HeaderSize() -> usize { 2 }
 
     pub fn parse(dat: &[u8]) -> ParseResult<Transform>
     {
